Recover callouts from CSS that marks values important
